fix(collector+fixer): tok_tgt in fp32 to prevent fp16 matmul overflow
Root cause of 100% val-acc with pred=label=[0,0,...] on our trained draft:
the earlier GPU-argmax patch (commit fd12229) computed the LM-head matmul
in fp16 over vocab=262144 × hidden=1536. Some output columns overflow to
Inf/NaN, which collapses torch.argmax to token 0 for every pair. The
resulting memmap dataset had tok_tgt = 0 everywhere; the draft then
trained to output 0 and the (wrong) labels matched 100%.

h_tgt.dat is fine (the overflow was only in the final vocab matmul, not
in the forward pass proper). Two fixes:

1. `collect_eagle_hidden_states_custom.py` (collector): force the LM-head
   matmul to fp32, `F.linear(hiddens_B[:, 1:].float(), lm_head_weight_gpu.float())`.
   Adds ~4-5 GB transient GPU memory; A100 handles it.

2. `fix_tok_tgt.py` (new utility): recompute tok_tgt.dat in-place from
   the existing h_tgt.dat in fp32 chunks. ~5-10 min on A100 for 7.7M
   pairs. No touch to h_tgt / e_in / fusion_L* / manifest. After
   running, the existing memmap dataset is usable for TTT retraining
   without re-collection.

Verification in the fixer: post-fix prints unique values + % zero in
first 10k. Hard-fails if % zero > 90% (indicates h_tgt or lm_head_weight
corruption, not the fp16 overflow pathology).

+++ conversion/collect_eagle_hidden_states_custom.py
@@ -489,10 +489,12 @@ def main():
         # hiddens_B: (B, max_N, hidden) ; fusion_list_B: [3 × (B, max_N, hidden)]
         embeds_B = model.embed_tokens(batched).to(MODEL_DTYPE) * embed_scale  # (B, max_N, hidden)
 
-        # LM head argmax on GPU (fp16 matmul — ~50ms for B=16 seq=512, vs ~16s
-        # if done on CPU per-sample). Only argmax is needed, not full logits, so
-        # fp16 precision is sufficient.
-        logits_B = F.linear(hiddens_B[:, 1:], lm_head_weight_gpu)  # (B, max_N-1, vocab)
+        # LM head argmax on GPU. MUST be fp32 — with vocab=262144 and
+        # hidden=1536, fp16 matmul overflows to Inf/NaN in ~some output
+        # columns, causing argmax to consistently land on token 0. Using
+        # fp32 keeps it correct at ~3 GB extra transient memory per batch
+        # (B=16, seq=512, vocab=262144 fp32 logits ≈ 8.5 GB).
+        logits_B = F.linear(hiddens_B[:, 1:].float(), lm_head_weight_gpu.float())
         tok_tgt_B = logits_B.argmax(dim=-1)  # (B, max_N-1) int64
 
         # Single GPU→CPU sync for the whole batch (amortized)
